backend/pyproject.toml
[tool.poetry]
name = "fastapi-brands"
version = "0.1.0"
description = "Backend hexagonal para gestión de marcas usando FastAPI"
authors = ["Tu Nombre <tu@email.com>"]
readme = "README.md"
packages = [{ include = "app", from = "src" }]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
pydantic = "^2.6.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.0"
injector = "^0.22.0"
psycopg2-binary = "^2.9.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
ruff = "^0.4.0"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
ipython = "^8.24.0"

[tool.poetry.scripts]
start = "app.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "B", "I"]
ignore = ["E501"]
exclude = ["tests"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = ["tests"]
